requirements.txt
git+https://github.com/pytorch/tnt.git@master
fire
numpy
torchvision
visdom
torchnet #我自己添加的